[Refactor] 파일 구조 변경 및 SlideManager에서 random Square Slide 생성하기
