+++ MyKeynote/MyKeynote/Model/Property/RGBColor.swift
@@ -12,20 +12,6 @@ struct RGBColor {
         self.blue = blue
     }
     
-    var hexadecimal: String {
-        var hexaString = "0x"
-        for rgb in [red, green, blue] {
-            let mod: Int = Int(rgb % 16)
-            let divisor: Int = Int(rgb / 16)
-            print(mod, divisor)
-
-            hexaString += (divisor >= 10 ? String(UnicodeScalar(divisor + 55)!) : String(divisor))
-
-            hexaString += (mod >= 10 ? String(UnicodeScalar(mod + 55)!) : String(mod))
-        }
-        return hexaString
-    }
-    
     static func randomNumber() -> UInt8 {
         return UInt8.random(in: 0...255)
     }
@@ -35,5 +21,17 @@ struct RGBColor {
     }
 }
 
+extension RGBColor {
+    var hexadecimal: String {
+        var hexaString = "0x"
+        for rgb in [red, green, blue] {
+            let mod: Int = Int(rgb % 16)
+            let divisor: Int = Int(rgb / 16)
+            hexaString += (divisor >= 10 ? String(UnicodeScalar(divisor + 55)!) : String(divisor))
+            hexaString += (mod >= 10 ? String(UnicodeScalar(mod + 55)!) : String(mod))
+        }
+        return hexaString
+    }
+}
