+++ MyKeynote/MyKeynote/Model/Slide/ImageSlide.swift
@@ -1,0 +1,8 @@
+//
+//  ImageSlide.swift
+//  MyKeynote
+//
+//  Created by 이수민 on 2023/07/20.
+//
+
+import Foundation
