+++ MyKeynote/MyKeynote/Model/Slide/SquareSlide.swift
