+++ MyKeynote/MyKeynote/Util/ConstantSize.swift
