+++ MyKeynote/MyKeynote/Util/StringExtension.swift
